# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "javascript_seminar_meeting",
    "version": "0.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:4200/bbbbridge/api/v1/bbb"
    }
  ],
  "paths": {
...
    "/administration/join": {
      "post": {
        "summary": "Create join-meeting url",
        "tags": [
          "administration"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "required": [
                  "fullName",
                  "meetingID",
                  "password"
                ],
                "properties": {
                  "fullName": {
                    "type": "string",
                    "description": "The full name that is to be used to identify this user to other conference attendees."
                  },
                  "meetingId": {
                    "type": "string",
                    "description": "The meeting ID that identifies the meeting you are attempting to join."
                  },
                  "password": {
                    "type": "string",
                    "description": "The password that this attendee is using. If the moderator password is supplied, he will be given moderator status (and the same for attendee password, etc)"
                  },
                  "createTime": {
                    "type": "string",
                    "description": "Third-party apps using the API can now pass createTime parameter (which was created in the create call), BigBlueButton will ensure it matches the ‘createTime’ for the session. If they differ, BigBlueButton will not proceed with the join request. This prevents a user from reusing their join URL for a subsequent session with the same meetingID."
                  },
                  "userID": {
                    "type": "string",
                    "description": "An identifier for this user that will help your application to identify which person this is. This user ID will be returned for this user in the getMeetingInfo API call so that you can check"
                  },
                  "webVoiceConf": {
                    "type": "string",
                    "description": "If you want to pass in a custom voice-extension when a user joins the voice conference using voip. This is useful if you want to collect more info in you Call Detail Records about the user joining the conference. You need to modify your /etc/asterisk/bbb-extensions.conf to handle this new extensions."
                  },
                  "configToken": {
                    "type": "string",
                    "description": "The token returned by a setConfigXML API call. This causes the BigBlueButton client to load the config.xml associated with the token (not the default config.xml)"
                  },
                  "defaultLayout": {
                    "type": "string",
                    "description": "The layout name to be loaded first when the application is loaded."
                  },
                  "avatarURL": {
                    "type": "string",
                    "description": "The link for the user’s avatar to be displayed when displayAvatar in config.xml is set to true (not yet implemented in the HTML5 client, see"
                  },
                  "redirect": {
                    "type": "string",
                    "description": "The default behaviour of the JOIN API is to redirect the browser to the Flash client when the JOIN call succeeds. There have been requests if it’s possible to embed the Flash client in a “container” page and that the client starts as a hidden DIV tag which becomes visible on the successful JOIN. Setting this variable to FALSE will not redirect the browser but returns an XML instead whether the JOIN call has succeeded or not. The third party app is responsible for displaying the client to the user."
                  },
                  "clientURL": {
                    "type": "string",
                    "description": "Some third party apps what to display their own custom client. These apps can pass the URL containing the custom client and when redirect is not set to false, the browser will get redirected to the value of clientURL."
                  },
                  "joinViaHtml5": {
                    "type": "string",
                    "description": "Set to “true” to force the HTML5 client to load for the user."
                  },
                  "guest": {
                    "type": "string",
                    "description": "Set to “true” to indicate that the user is a guest, otherwise do NOT send this parameter."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/administration/end": {
      "post": {
        "summary": "Creates end-meeting url",
        "tags": [
          "administration"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "required": [
                  "meetingId",
                  "moderatorPW"
                ],
                "properties": {
                  "meetingId": {
                    "type": "string",
                    "description": "The meeting ID that identifies the meeting you are attempting to end."
                  },
                  "moderatorPW": {
                    "type": "string",
                    "description": "The moderator password for this meeting. You can not end a meeting using the attendee password."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "summary": "Ends a meeting",
        "tags": [
          "administration"
        ],
        "parameters": [
          {
            "name": "meetingId",
            "description": "The meeting ID that identifies the meeting you are attempting to end.",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "description": "The moderator password for this meeting. You can not end a meeting using the attendee password.",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}